+++ CHANGES.md
@@ -21,6 +21,7 @@
 * Lib: fix various Dom_html bindings (submitEvent, mediaQueryListEvent, pointerEvent, element, inputElement, tableElement types and deprecations)
 * Lib: fix `numberList` type in Dom_svg to use `number_t`
 * Lib: fix `eventSource.url` type to use `js_string t`
+* Lib: fix `file.lastModifiedDate` type to use `Js.date t`, add `lastModified`
 * Compiler: fix purity of comparison functions (again) (#2092)
 * Compiler: fix inlining (#2107)
 * Compiler: allow arrow functions in for loops